#11 부분수정된 파일 커밋

+++ src/org/iclass/vo/CommunityComments.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @Builder
 public class CommunityComments {
 	private int idx;
-	private int mref;
+	private long mref;
 	private String writer;
 	private String content;
 	private Timestamp createdAt;
